Rename folder name integration to integrations in image2image and animation
