+++ .github/workflows/docker/compose/animation-compose.yaml
@@ -9,9 +9,9 @@ services:
     image: ${REGISTRY:-opea}/animation:${TAG:-latest}
   wav2lip:
     build:
-      dockerfile: comps/animation/src/integration/dependency/Dockerfile
+      dockerfile: comps/animation/src/integrations/dependency/Dockerfile
     image: ${REGISTRY:-opea}/wav2lip:${TAG:-latest}
   wav2lip-gaudi:
     build:
-      dockerfile: comps/animation/src/integration/dependency/Dockerfile.intel_hpu
+      dockerfile: comps/animation/src/integrations/dependency/Dockerfile.intel_hpu
     image: ${REGISTRY:-opea}/wav2lip-gaudi:${TAG:-latest}
