+++ .github/workflows/scripts/get_test_matrix.sh
@@ -42,7 +42,7 @@ function find_test_1() {
                     changed_integrations=$(printf '%s\n' "${changed_files[@]}"| grep ${service_path} | grep -E 'integrations' | cut -d'/' -f$((n+2)) | cut -d'.' -f1 | sort -u)  || true
                     for integration in ${changed_integrations}; do
                         # Accurate matching test scripts
-                        # find_test=$(find ./tests -type f \( -name test_${service_name}_${integration}.sh -o -name test_${service_name}_${integration}_on_*.sh \)) || true
+                        # find_test=$(find ./tests -type f \( -name test_${service_name}_${integrations}.sh -o -name test_${service_name}_${integrations}_on_*.sh \)) || true
                         # Fuzzy matching test scripts, for example, llms/src/text-generation/integrations/opea.py match several tests.
                         find_test=$(find ./tests -type f -name test_${service_name}_${integration}*.sh) || true
                         if [ "$find_test" ]; then
