+++ comps/animation/src/integrations/dependency/Dockerfile.intel_hpu
@@ -19,7 +19,7 @@ RUN rm -rf /var/lib/apt/lists/*
 # Install GenAIComps
 RUN mkdir -p /home/user/comps
 COPY comps /home/user/comps
-COPY comps/animation/src/integration/dependency/entrypoint.sh /usr/local/bin/entrypoint.sh
+COPY comps/animation/src/integrations/dependency/entrypoint.sh /usr/local/bin/entrypoint.sh
 
 # Install ffmpeg with x264 software codec
 RUN git clone https://github.com/FFmpeg/FFmpeg.git /home/user/comps/animation/FFmpeg
