+++ comps/animation/src/integrations/dependency/download_ckpts.sh
