+++ comps/image2image/src/integrations/opea_image2image_native.py
